feat: can driver on stm32h7 & small edit on flash write

+++ include/libparams/flash_driver.h
@@ -51,7 +51,7 @@ size_t flashRead(uint8_t* data, size_t offset, size_t bytes_to_read);
  * @return Info about the flash memory
  */
 uint16_t flashGetNumberOfPages();
-uint16_t flashGetPageSize();
+uint32_t flashGetPageSize();
 
 #ifdef __cplusplus
 }
